+++ package.json
@@ -6,18 +6,28 @@
     "container",
     "server"
   ],
+  "eslintConfig": {
+    "root": true
+  },
   "scripts": {
     "build": "npm run -ws build --if-present",
-    "lint": "npm run -ws lint --if-present",
-    "test": "npm run -ws test --if-present",
+    "lint": "eslint .",
+    "test": "jest",
+    "test-coverage": "jest --coverage",
     "update": "npx npm-check-updates -u -ws && npm update -ws"
   },
   "type": "module",
   "devDependencies": {
+    "@types/jest": "^29.5.8",
+    "@types/node": "^20.9.0",
     "@typescript-eslint/eslint-plugin": "^6.10.0",
     "@typescript-eslint/parser": "^6.10.0",
     "eslint": "^8.53.0",
+    "jest": "^29.7.0",
     "npm-check-updates": "^16.14.6",
+    "ts-jest": "^29.1.1",
+    "ts-node": "^10.9.1",
+    "tsx": "^3.14.0",
     "typescript": "^5.2.2"
   }
 }
