+++ server/package.json
@@ -8,13 +8,11 @@
     "versatiles-server": "node dist/index.js"
   },
   "scripts": {
-    "check": "npm run lint && npm run build && npm run test",
+    "check": "npm run lint && npm run test && npm run build",
     "build": "rm -rf dist && tsc -p tsconfig.build.json",
     "doc": "npm run -w tsdoc2readme update_readme container/src/index.ts '# API'",
-    "lint": "eslint .",
     "start": "tsx src/index.ts",
-    "test": "NODE_OPTIONS=--experimental-vm-modules jest",
-    "test-coverage": "NODE_OPTIONS=--experimental-vm-modules jest --coverage"
+    "test": "jest"
   },
   "type": "module",
   "author": "yetzt <node@yetzt.me>, Michael Kreil <versatiles@michael-kreil.de>",
@@ -40,17 +38,6 @@
     "vectortiles",
     "versatiles"
   ],
-  "devDependencies": {
-    "@types/jest": "^29.5.8",
-    "@types/node": "^20.9.0",
-    "jest": "^29.7.0",
-    "jest-ts-webcompat-resolver": "^1.0.0",
-    "tar-stream": "^3.1.6",
-    "ts-jest": "^29.1.1",
-    "ts-node": "^10.9.1",
-    "tsx": "^3.14.0",
-    "typescript": "^5.2.2"
-  },
   "dependencies": {
     "@versatiles/container": "^1.0.0",
     "@versatiles/style": "^3.4.1",
