+++ server/tsconfig.build.json
@@ -1,8 +1,11 @@
 {
-	"extends": "./tsconfig.json",
+	"extends": "../tsconfig.json",
 	"compilerOptions": {
 		"outDir": "dist",
 	},
+	"include": [
+		"src/**/*.ts"
+	],
 	"exclude": [
 		"src/**/*.test.ts"
 	]
